+++ common/appointment/util.ts
@@ -25,3 +25,5 @@ export function getAppointmentForNotification(appointment: Appointment, original
         original: original ? getAppointmentForNotification(original) : undefined,
     };
 }
+
+export class AuthorizationError extends Error {}
